Put parse_definition related funcs into own module
